+ login

+++ app/routers/blog.py
@@ -1,7 +1,7 @@
 import sqlite3
-from fastapi import APIRouter, status, HTTPException
+from fastapi import APIRouter, status, HTTPException, Depends
 from utils.others import get_dict
-from .. import schema
+from .. import schema, oauth2
 
 router = APIRouter(
     prefix="/blogs",
@@ -10,23 +10,29 @@ router = APIRouter(
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
-def create_blog(data: schema.BlogCreate):
+def create_blog(data: schema.BlogCreate, current_user: schema.TokenData = Depends(oauth2.get_current_user)):
     """
     Creates a blog
     """
-    with sqlite3.connect("acm.db") as db:
-        cur = db.cursor()
-        try:
-            cur.execute(
-                "INSERT INTO blogs VALUES(:title, :description, :date, :author, :image_url, :link)",
-                data.model_dump(),
-            )
-        except sqlite3.IntegrityError:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Title Already Exists"
-            )
-        db.commit()
-    return {"message": "Blog Added."}
+    if current_user.admin:
+        with sqlite3.connect("acm.db") as db:
+            cur = db.cursor()
+            try:
+                cur.execute(
+                    "INSERT INTO blogs VALUES(:title, :description, :date, :author, :image_url, :link)",
+                    data.model_dump(),
+                )
+            except sqlite3.IntegrityError:
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST, detail="Title Already Exists"
+                )
+            db.commit()
+        return {"message": "Blog Added."}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You don't have permission to perform this action."
+        )
 
 
 @router.get("/")
@@ -42,39 +48,51 @@ def blogs():
 
 
 @router.patch("/")
-def update_blog(data: schema.BlogUpdate):
-    title = data.title
-    with sqlite3.connect("acm.db") as db:
-        cur = db.cursor()
-        cur.execute("SELECT * FROM blogs WHERE title = ?", (title,))
-        if not cur.fetchone():
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Blog Not Found"
-            )
-    _data = data.model_dump()
-    _data = dict(filter(lambda x: x[1] is not None, _data.items()))
-    _data.pop("title")
-    _data = dict(map(lambda x: (x[0].replace("new_", ""), x[1]), _data.items()))
-    with sqlite3.connect("acm.db") as db:
-        cur = db.cursor()
-        for k, v in _data.items():
-            cur.execute(f"UPDATE blogs SET {k} = ? WHERE title = ?", (v, title))
-            if k == "title":
-                title = v
-        db.commit()
-    return {"message": "Updated Successfully"}
+def update_blog(data: schema.BlogUpdate, current_user: schema.TokenData = Depends(oauth2.get_current_user)):
+    if current_user.admin:
+        title = data.title
+        with sqlite3.connect("acm.db") as db:
+            cur = db.cursor()
+            cur.execute("SELECT * FROM blogs WHERE title = ?", (title,))
+            if not cur.fetchone():
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST, detail="Blog Not Found"
+                )
+        _data = data.model_dump()
+        _data = dict(filter(lambda x: x[1] is not None, _data.items()))
+        _data.pop("title")
+        _data = dict(map(lambda x: (x[0].replace("new_", ""), x[1]), _data.items()))
+        with sqlite3.connect("acm.db") as db:
+            cur = db.cursor()
+            for k, v in _data.items():
+                cur.execute(f"UPDATE blogs SET {k} = ? WHERE title = ?", (v, title))
+                if k == "title":
+                    title = v
+            db.commit()
+        return {"message": "Updated Successfully"}
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You don't have permission to perform this action."
+        )
 
 
 @router.delete("/", status_code=status.HTTP_204_NO_CONTENT)
-def delete_blog(data: schema.BlogDelete):
+def delete_blog(data: schema.BlogDelete, current_user: schema.TokenData = Depends(oauth2.get_current_user)):
     """
     Deletes the blog
     """
-    with sqlite3.connect("acm.db") as db:
-        cur = db.cursor()
-        cur.execute("SELECT * FROM blogs WHERE title = ?", (data.title,))
-        if not cur.fetchone():
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST, detail="Blog Not Found"
-            )
-        cur.execute("DELETE FROM blogs WHERE title = ?", (data.title,))
+    if current_user.admin:
+        with sqlite3.connect("acm.db") as db:
+            cur = db.cursor()
+            cur.execute("SELECT * FROM blogs WHERE title = ?", (data.title,))
+            if not cur.fetchone():
+                raise HTTPException(
+                    status_code=status.HTTP_400_BAD_REQUEST, detail="Blog Not Found"
+                )
+            cur.execute("DELETE FROM blogs WHERE title = ?", (data.title,))
+    else:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You don't have permission to perform this action."
+        )
